rag-agents/requirements.txt
# This file is used for the dockerisation of the agent
litellm
google-adk
#a2a-sdk
langchain-core
#langchain-litellm
langchain-openai